GAN-CosineSimility/requirements.txt
torch>=2.0.0
numpy>=1.22.0
scikit-learn>=1.2.0
PyYAML>=6.0
tensorboard>=2.12.0
tqdm>=4.65.0
open3d>=0.17.0
joblib>=1.2.0
pandas>=1.5.3